requirements.txt
certifi==2019.9.11
chardet==3.0.4
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.6
django-crispy-forms==1.7.2
gunicorn==19.9.0
idna==2.8
Pillow==6.2.0
psycopg2==2.8.3
pyperclip==1.7.0
pytz==2019.3
requests==2.22.0
Send2Trash==1.5.0
sorl-thumbnail==12.5.0
sqlparse==0.3.0
static3==0.7.0
stripe==2.37.2
urllib3==1.25.6
virtualenv==16.7.6
whitenoise==4.1.4
